API

API

FamaMacBeth.fmb — Method.

fmb(X, y, g) -> FMBResp()

Runs a Fama-MacBeth (1973) panel regression by first conducting OLS of y on X per group indicated by g, and subsequently computes the mean and standard error using the resulting coefficients. An intercept is not added to X.

source

FamaMacBeth.nobs — Method.

nobs(m::FMBResp) -> Vector{Int}

Returns the number of observations, per group as identified by g. See also group(m::FMBResp).

source

FamaMacBeth.tstat — Method.

tstat(m::FMBResp) -> Vector{}

The t-statistic per coefficient whether the coefficient equals zero or not.

source

StatsBase.coef — Method.

coef(m::FMBResp) -> Vector{}

The estimated coefficients.

source

StatsBase.stderror — Method.

stderror(m::FMBResp) -> Vector{}

The standard error around the coefficients.

source

FamaMacBeth.group — Method.

group(m::FMBResp)

Returns the group labels used.

source