API

API

FamaMacBeth.adjr2 — Method.

adjr2(m::FMBModel) -> Float64

The adjusted R2.

source

FamaMacBeth.fmb — Method.

fmb(X, y, g) -> FMBModel()

Short-hand for fit(FMBModel(X,y,g))

source

FamaMacBeth.group — Method.

group(m::FMBResp)
group(m::FMBModel)

Returns the group labels used.

source

FamaMacBeth.nobs — Method.

nobs(m::FMBModel) -> Vector{Int}

Returns the number of observations, per group as identified by g. See also group(m::FMBResp).

source

FamaMacBeth.predict — Method.

predict(m::FMBModel) -> Vector
predict(m::FMBModel, X) -> Vector

Predicts the outcome using the estimated coefficients and provided independent variables. If independent variables are not provided, uses those from the model.

source

FamaMacBeth.r2 — Method.

r2(m::FMBModel) -> Float64

The un-adjusted R2.

source

FamaMacBeth.residuals — Method.

residuals(m::FMBModel) -> Vector

source

FamaMacBeth.tstat — Method.

tstat(m::FMBModel) -> Vector
tstat(m::FMBModel) -> Vector

The t-statistic per coefficient whether the coefficient equals zero or not.

source

StatsBase.coef — Method.

coef(m::FMBResp) -> Vector
coef(m::FMBModel) -> Vector

The estimated coefficients.

source

StatsBase.stderror — Method.

stderror(m::FMBResp) -> Vector
stderror(m::FMBModel) -> Vector

The standard error around the estimated coefficients.

source

FamaMacBeth.FMBModel — Type.

FMBModel(X,y,g,resp::FMBResp) <: StatsBase.RegressionModel

source

FamaMacBeth.FMBModel — Method.

FMBModel(X,y,g) -> FMBModel()

source

FamaMacBeth.fit — Method.

fit(m::FMBModel) -> FMBModel()

Runs a Fama-MacBeth (1973) panel regression by first conducting OLS of y on X per group indicated by g, and subsequently computes the mean and standard error using the resulting coefficients. An intercept is not added to X.

source

FamaMacBeth.vcov — Method.

vcov(m::FMBResp) -> Matrix
vcov(m::FMBModel) -> Matrix

Variance-covariance matrix of the estimated coefficients.

source